requirements.txt
streamlit
pandas
sqlalchemy
openai
python-dotenv